fix(sonar): triage 120 open issues + 3 hotspots — fix 80, mark 40, harden ReDoS
SonarQube sweep for abundent:jclaw (branch sonar-fixes). Clears the single
BLOCKER driving the red quality gate (new_reliability_rating=2) and works the
safe idiomatic backlog; deliberate/idiomatic items are marked FP/accept in Sonar.

Gate driver:
- S2095 (BLOCKER): VoiceController leaked the native VoiceVad when session
  wiring threw before the VoiceSession took ownership. A handed-off flag +
  conditional close() in a finally now releases it on the init-failure path.

Code fixes (80 issues, 26 files):
- VoiceController: S2095 leak, S1192 frame-token constants, S135 loop
  break/continue, S7158 isEmpty, S1481/S7467 unnamed patterns, S1612 method ref.
- TtsText: S5852 ReDoS hotspot — collapsed the table-separator regex to one
  anchored character class (language-identical, no overlapping quantifiers).
- Backend batch: S3358, S1192, S1905, S1121, S7467, S135, S6126, S7158, S6885.
- Frontend batch: S7781, S7758, S6582, S3358.
- Test batch: S1128, S7467, S6213.

Hotspots: TtsText S5852 FIXED; useReadAloud S5852 + TtsJvmEngine S4036 SAFE.
Sonar marks: 8 false-positive (S3077, S2160, S6819, S2583 — framework/idiom +
ConcurrentMap.remove nullability mis-model), 32 accepted (S3776 x10, S899, S2445,
S107, S6541, S1141, S5443, S1452, S5778, S5976, S6548, S4138, S7764).

Validated: ./jclaw.sh test (port 9885) — backend 408 classes, frontend 1111
tests, stylelint/lint/typecheck all green.

+++ app/memory/JpaMemoryStore.java
@@ -494,7 +494,7 @@ public class JpaMemoryStore implements MemoryStore {
     private void generateAndStoreEmbedding(Memory memory) {
         var embedding = generateEmbedding(memory.text);
         if (embedding != null) {
-            storeEmbeddingSql((Long) memory.id, embedding);
+            storeEmbeddingSql(memory.id, embedding);
         }
     }
